Mutations without navigation
So far all of our mutations (the times we change data) have used forms
that navigate, creating new entries in the history stack. While these
user flows are common, it's equally as common to want to change data
without causing a navigation.

For these cases, we have the useFetcher hook. It allows us to
communicate with loaders and actions without causing a navigation.

The ★ button on the contact page makes sense for this. We aren't
creating or deleting a new record, we don't want to change pages, we
simply want to change the data on the page we're looking at.