Optimistic UI
You probably noticed the app felt kind of unresponsive when we clicked
the favorite button from the last section. Once again, we added some
network latency because you're going to have it in the real world!

To give the user some feedback, we could put the star into a loading
state with fetcher.state (a lot like navigation.state from before), but
we can do something even better this time. We can use a strategy called
"optimistic UI"

The fetcher knows the form data being submitted to the action, so it's
available to you on fetcher.formData. We'll use that to immediately
update the star's state, even though the network hasn't finished. If the
update eventually fails, the UI will revert to the real data.