Pathless routes
The last error page we saw would be better if it rendered inside the
root outlet, instead of the whole page. In fact, every error in all of
our child routes would be better in the outlet, then the user has more
options than hitting refresh.

We could add the error element to every one of the child routes but,
since it's all the same error page, this isn't recommended.

There's a cleaner way. Routes can be used without a path, which lets
them participate in the UI layout without requiring new path segments in
the URL.